api通信を行うクラスをinterface化して処理を切り分け、カプセル化した。
