+++ app/src/main/kotlin/jp/co/yumemi/android/code_check/di/AppModule.kt
@@ -5,6 +5,7 @@ import dagger.Provides
 import dagger.hilt.InstallIn
 import dagger.hilt.components.SingletonComponent
 import jp.co.yumemi.android.code_check.api.Api
+import jp.co.yumemi.android.code_check.api.ApiImpl
 import jp.co.yumemi.android.code_check.repository.SearchRepository
 import jp.co.yumemi.android.code_check.repository.SearchRepositoryImpl
 import javax.inject.Singleton
@@ -14,9 +15,9 @@ import javax.inject.Singleton
 object AppModule {
     @Provides
     @Singleton
-    fun provideSearchRepository(): SearchRepository = SearchRepositoryImpl(api = Api())
+    fun provideSearchRepository(api: Api): SearchRepository = SearchRepositoryImpl(api = api)
 
     @Provides
     @Singleton
-    fun provideApi(): Api = Api()
+    fun provideApi(): Api = ApiImpl()
 }
